refactor: 주문 테이블과 단체테이블 패키지 분리
