+++ src/main/java/kitchenpos/tablegroup/controller/TableGroupRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.ordertable.controller;
+package kitchenpos.tablegroup.controller;
 
-import kitchenpos.ordertable.application.TableGroupService;
-import kitchenpos.ordertable.controller.dto.TableGroupCreateRequest;
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.controller.dto.TableGroupCreateRequest;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
