+++ src/main/java/kitchenpos/tablegroup/controller/dto/TableGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.controller.dto;
+package kitchenpos.tablegroup.controller.dto;
 
 import java.util.List;
