+++ src/main/java/kitchenpos/tablegroup/domain/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.ordertable.domain.repository;
+package kitchenpos.tablegroup.domain.repository;
 
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
