+++ src/main/java/kitchenpos/tablegroup/exception/InvalidTableGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.exception;
+package kitchenpos.tablegroup.exception;
 
 public class InvalidTableGroupException extends RuntimeException {
     public InvalidTableGroupException(final String message) {
