+++ src/main/java/kitchenpos/tablegroup/exception/InvalidTableGroupUngroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.exception;
+package kitchenpos.tablegroup.exception;
 
 public class InvalidTableGroupUngroupException extends RuntimeException {
     public InvalidTableGroupUngroupException(final String message) {
