+++ src/main/java/kitchenpos/tablegroup/exception/NotExistTableGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.exception;
+package kitchenpos.tablegroup.exception;
 
 public class NotExistTableGroupException extends RuntimeException {
     public NotExistTableGroupException(final String message) {
