+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -1,13 +1,13 @@
 package kitchenpos.application;
 
-import kitchenpos.ordertable.controller.dto.TableGroupCreateRequest;
+import kitchenpos.tablegroup.controller.dto.TableGroupCreateRequest;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.TableGroup;
-import kitchenpos.ordertable.exception.InvalidTableGroupException;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.exception.InvalidTableGroupException;
 import kitchenpos.ordertable.exception.NotExistOrderTable;
 import kitchenpos.order.domain.repository.OrderRepository;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
-import kitchenpos.ordertable.application.TableGroupService;
+import kitchenpos.tablegroup.application.TableGroupService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
