+++ src/test/java/kitchenpos/fixture/OrderTableFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class OrderTableFixture {
